requirements.txt
# Smart Home AI Agent - Python Dependencies
# Production-grade requirements with pinned versions

# ===== Core Framework =====
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
streamlit==1.28.1
python-dotenv==1.0.0

# ===== MQTT & Networking =====
paho-mqtt==1.6.1
requests==2.31.0

# ===== Data Science & ML =====
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# ===== Time Series Forecasting =====
prophet==1.1.5
statsmodels==0.14.0

# ===== Deep Learning (Optional) =====
tensorflow==2.14.0
keras==2.14.0

# ===== Data Loading =====
kaggle==1.5.13

# ===== Visualization =====
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# ===== LLM Integration =====
openai==1.3.7

# ===== Logging & Monitoring =====
python-json-logger==2.0.7

# ===== Development & Testing =====
pytest==7.4.3
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# ===== Database (Optional) =====
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# ===== Async Support =====
aiohttp==3.9.1
asyncio==3.4.3

# ===== Security =====
cryptography==41.0.7
